code/requirements.txt
openai>=2.33.0
pyserial>=3.5
luma.oled>=3.14.0
RPi.GPIO>=0.7.1
